Replace the Interaction_Policy trait with per-model methods
Inline the canFeature computation into the User and Blog models and
only merge with an interaction policy when one is explicitly set. The
trait's implicit requirements on the using class (inherited accessor,
actor methods) made it too easy to use incorrectly.
